Renames tpm -> transition and minor fixes after review
